chore(IT Wallet): [SIW-349] Refactor state structure (#4892)
## Short description
This PR proposes a range of enhancements to the store structure.

## List of changes proposed in this pull request
- Adds the `itw` prefix to saga/actions/reducers as naming convention
for IT wallet related files. Introduces the "Saga" suffix as a naming
convention for sagas.
- For each saga file, includes a `watcher` function to spawn related
tasks, enhancing main saga organization and navigation.
- Forks the mentioned watcher in the index file, further forked in the
startup saga;
- Moves `Activation` related handlers/Actions to dedicated
`itwActivationSaga` and `itwActivationAction` files.
- Moves `WIA`, `Lifecycle`, `Activation` actions to dedicated files;
- Updates imports to align with changes.

Please note that CIE elements are intentionally omitted, as they have
been directly copied from the existing CIE implementation.

## How to test
Test the whole wallet activation process from scratch and navigate to
the PID preview screen. Also try to abort it.
