+++ ts/features/it-wallet/screens/ItwHomeScreen.tsx
@@ -18,19 +18,19 @@ import { VSpacer } from "../../../components/core/spacer/Spacer";
 import { ITW_ROUTES } from "../navigation/ItwRoutes";
 import ButtonLink from "../../../components/ui/ButtonLink";
 import { useItwResetFlow } from "../hooks/useItwResetFlow";
-import { itwLifecycleIsOperationalSelector } from "../store/reducers/itwLifecycle";
-import { ItwCredentialsPidSelector } from "../store/reducers/itwCredentials";
 import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
 import PidCredential from "../components/PidCredential";
 import { IOStackNavigationProp } from "../../../navigation/params/AppParamsList";
 import { ItwParamsList } from "../navigation/ItwParamsList";
-import { itwDecodePid } from "../store/actions/credentials";
 import LoadingSpinnerOverlay from "../../../components/LoadingSpinnerOverlay";
-import { ItwDecodedPidPotSelector } from "../store/reducers/itwPidDecode";
 import { ItWalletError } from "../utils/errors/itwErrors";
 import { mapRequirementsError } from "../utils/errors/itwErrorsMapping";
 import { InfoScreenComponent } from "../../fci/components/InfoScreenComponent";
 import FooterWithButtons from "../../../components/ui/FooterWithButtons";
+import { itwLifecycleIsOperationalSelector } from "../store/reducers/itwLifecycleReducer";
+import { ItwCredentialsPidSelector } from "../store/reducers/itwCredentialsReducer";
+import { ItwDecodedPidPotSelector } from "../store/reducers/itwPidDecodeReducer";
+import { itwDecodePid } from "../store/actions/itwCredentialsActions";
 
 const contextualHelpMarkdown: ContextualHelpPropsMarkdown = {
   title: "wallet.contextualHelpTitle",
