+++ ts/features/it-wallet/screens/issuing/ItwPidRequestScreen.tsx
@@ -13,15 +13,15 @@ import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender";
 import { ItwParamsList } from "../../navigation/ItwParamsList";
 import { IOStackNavigationProp } from "../../../../navigation/params/AppParamsList";
 import { useIODispatch, useIOSelector } from "../../../../store/hooks";
-import { itwPid } from "../../store/actions/credentials";
-import { itwPidSelector } from "../../store/reducers/itwPid";
 import { InfoScreenComponent } from "../../../fci/components/InfoScreenComponent";
 import { ItWalletError } from "../../utils/errors/itwErrors";
 import { mapRequirementsError } from "../../utils/errors/itwErrorsMapping";
 import { Pictogram } from "../../../../components/core/pictograms";
 import ItwLoadingSpinnerOverlay from "../../components/ItwLoadingSpinnerOverlay";
 import { ITW_ROUTES } from "../../navigation/ItwRoutes";
-import { itwActivationStop } from "../../store/actions";
+import { itwPidSelector } from "../../store/reducers/itwPidReducer";
+import { itwPid } from "../../store/actions/itwCredentialsActions";
+import { itwActivationStop } from "../../store/actions/itwActivationActions";
 
 /**
  * ItwPidRequestScreen's navigation params.
