+++ ts/features/it-wallet/store/actions/index.ts
@@ -1,71 +1,15 @@
-import {
-  ActionType,
-  createStandardAction,
-  createAsyncAction
-} from "typesafe-actions";
-import { ItWalletError } from "../../utils/errors/itwErrors";
-import { itwCredentialsActions } from "./credentials";
-import { ItwCieAuthenticationActions } from "./cie";
-
-/**
- * Start ITW activation
- */
-export const itwActivationStart = createStandardAction(
-  "ITW_ACTIVATION_START"
-)<void>();
-
-/**
- * Stop ITW activation
- */
-export const itwActivationStop = createStandardAction(
-  "ITW_ACTIVATION_STOP"
-)<void>();
-
-/**
- * Complete ITW activation
- */
-export const itwActivationCompleted = createStandardAction(
-  "ITW_ACTIVATION_COMPLETED"
-)<void>();
-
-/**
- * Start ITW wallet instance attestation request.
- */
-export const itwWiaRequest = createAsyncAction(
-  "ITW_REQUIREMENTS_REQUEST",
-  "ITW_REQUIREMENTS_SUCCESS",
-  "ITW_REQUIREMENTS_FAILURE"
-)<void, string, ItWalletError>();
-
-/**
- * Actions which sets the wallet lifecycle status to operational.
- */
-export const itwLifecycleOperational = createStandardAction(
-  "ITW_LIFECYCLE_OPERATIONAL"
-)<void>();
-
-/**
- * Actions which sets the wallet lifecycle status to valid.
- */
-export const itwLifecycleValid = createStandardAction(
-  "ITW_LIFECYCLE_VALID"
-)<void>();
-
-/**
- * Actions which sets the wallet lifecycle status to deactivated.
- */
-export const itwLifecycleDeactivated = createStandardAction(
-  "ITW_LIFECYCLE_DEACTIVATED"
-)<void>();
+import { ItwCredentialsActions } from "./itwCredentialsActions";
+import { ItwCieAuthenticationActions } from "./itwCieActions";
+import { ItwActivationActions } from "./itwActivationActions";
+import { ItwWiaActions } from "./itwWiaActions";
+import { ItwLifecycleActions } from "./itwLifecycleActions";
 
 /**
  * Action types for the IT Wallet feature
  */
 export type ItWalletActions =
-  | ActionType<typeof itwActivationStart>
-  | ActionType<typeof itwWiaRequest>
-  | ActionType<typeof itwLifecycleOperational>
-  | ActionType<typeof itwLifecycleValid>
-  | ActionType<typeof itwLifecycleDeactivated>
-  | itwCredentialsActions
+  | ItwActivationActions
+  | ItwWiaActions
+  | ItwLifecycleActions
+  | ItwCredentialsActions
   | ItwCieAuthenticationActions;
