+++ ts/features/it-wallet/store/actions/itwWiaActions.ts
@@ -1,0 +1,16 @@
+import { ActionType, createAsyncAction } from "typesafe-actions";
+import { ItWalletError } from "../../utils/errors/itwErrors";
+
+/**
+ * Start ITW wallet instance attestation request.
+ */
+export const itwWiaRequest = createAsyncAction(
+  "ITW_REQUIREMENTS_REQUEST",
+  "ITW_REQUIREMENTS_SUCCESS",
+  "ITW_REQUIREMENTS_FAILURE"
+)<void, string, ItWalletError>();
+
+/**
+ * Type for WIA related actions.
+ */
+export type ItwWiaActions = ActionType<typeof itwWiaRequest>;
