+++ ts/features/it-wallet/store/reducers/index.ts
@@ -3,12 +3,12 @@ import { PersistConfig, PersistPartial, persistReducer } from "redux-persist";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { Action } from "../../../../store/actions/types";
 import createCredentialsKeychain from "../storages/credentialsKeychain";
-import itwCieReducer, { ItwCieState } from "./cie";
-import itwWia, { ItwWIAState } from "./itwWia";
-import itwCredentials, { ItwCredentialsState } from "./itwCredentials";
-import itwLifeCycle, { ItwLifecycleState } from "./itwLifecycle";
-import itwPid, { ItwPidState } from "./itwPid";
-import itwDecodedPid, { ItwDecodedPidState } from "./itwPidDecode";
+import itwCieReducer, { ItwCieState } from "./itwCieReducer";
+import itwWia, { ItwWIAState } from "./itwWiaReducer";
+import itwCredentials, { ItwCredentialsState } from "./itwCredentialsReducer";
+import itwLifeCycle, { ItwLifecycleState } from "./itwLifecycleReducer";
+import itwPid, { ItwPidState } from "./itwPidReducer";
+import itwDecodedPid, { ItwDecodedPidState } from "./itwPidDecodeReducer";
 
 const CURRENT_REDUX_ITW_STORE_VERSION = 1;
